requirements.txt
langchain
PyPDF2
pypdf
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
huggingface-hub
InstructorEmbedding
sentence-transformers
flask